доавил метрики

+++ OTDemo.Solution/Weather.Libs/Models/WeatherData.cs
@@ -4,7 +4,7 @@ public class WeatherData
 {
     public string? Id { get; set; }
     public string? City { get; set; }
-    public double Temperature { get; set; }
+    public int Temperature { get; set; }
     public int Humidity { get; set; }
     public string? Condition { get; set; }
     public DateTime Date { get; set; }
